chore: Move playlist utils over
- `getPlaylistArtwork` will be deprecated as we handle the artwork logic within SQL / updated api functions.
