+++ mobile/src/data/playlist/queries.ts
@@ -3,10 +3,10 @@ import { useTranslation } from "react-i18next";
 
 import type { playlists } from "~/db/schema";
 
-import { sanitizePlaylistName } from "~/api/playlist.utils";
 import { Resynchronize } from "~/stores/Playback/actions";
 import { queries as q } from "~/queries/keyStore";
 import { createPlaylist, deletePlaylist, updatePlaylist } from "./api";
+import { sanitizePlaylistName } from "./utils";
 
 import { wait } from "~/utils/promise";
