+++ mobile/src/navigation/screens/playlists/components/ModifyViewBase.tsx
@@ -13,9 +13,9 @@ import { Cancel } from "~/resources/icons/Cancel";
 import { CheckCircle } from "~/resources/icons/CheckCircle";
 import { DragHandle } from "~/resources/icons/DragHandle";
 import { getArtistsString } from "~/api/artist.utils";
-import { sanitizePlaylistName } from "~/api/playlist.utils";
 import { TrackList, getTrackArtwork } from "~/api/track.utils";
 import { useDeletePlaylist } from "~/data/playlist/queries";
+import { sanitizePlaylistName } from "~/data/playlist/utils";
 
 import { useFloatingContent } from "~/navigation/hooks/useFloatingContent";
 import { ContentPlaceholder } from "~/navigation/components/Placeholder";
